requirements.txt
asn1crypto==1.4.0
azure-appconfiguration==1.1.1
azure-common==1.1.26
azure-core==1.9.0
azure-functions==1.5.0
azure-functions-durable==1.0.0b11
azure-keyvault==4.1.0
azure-keyvault-certificates==4.2.1
azure-keyvault-keys==4.3.1
azure-keyvault-secrets==4.2.0
azure-storage-blob==12.6.0
boto3==1.16.32
botocore==1.19.32
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cryptography==3.2.1
idna==2.10
isodate==0.6.0
jmespath==0.10.0
msrest==0.6.19
numpy==1.19.4
oauthlib==3.1.0
oscrypto==1.2.1
pandas==1.1.5
pyarrow==0.17.1
pycparser==2.20
pycryptodomex==3.9.9
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2020.4
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
snowflake-connector-python==2.3.7
twilio==6.50.0
urllib3==1.25.11
